fix the bugs where gride_size type in voxelize custom op is set as float but supposed to be int

+++ deploy/centerpoint/cpp/custom_ops/voxelize_op.cu
@@ -123,8 +123,8 @@ assign_voxels_kernel(const T *points, const T_int *points_to_grid_idx,
 template <typename T, typename T_int>
 __global__ void assign_coords_kernel(
     const T_int *grid_idx_to_voxel_idx, const T_int *num_points_in_grid,
-    const int num_grids, const float grid_size_x, const float grid_size_y,
-    const float grid_size_z, T *coords, T *num_points_per_voxel) {
+    const int num_grids, const int grid_size_x, const int grid_size_y,
+    const int grid_size_z, T *coords, T *num_points_per_voxel) {
   int64_t grid_idx = blockIdx.x * blockDim.x + threadIdx.x;
   if (grid_idx > num_grids || grid_idx == num_grids) {
     return;
